feat: wildcard update
