+++ Audacia.OrToolsPlayground/Examples/PickFplTeam/Models/PickFplTeamModel.cs
@@ -35,19 +35,4 @@ namespace Audacia.OrToolsPlayground.Examples.PickFplTeam.Models
 
         public PickFplTeamOptions Options { get; }
     }
-
-    public class PickFplTeamOptions
-    {
-        public int MaxPlayersPerTeam { get; set; }
-
-        public int Budget { get; set; }
-
-        public int NumberGoalkeepers { get; set; }
-
-        public int NumberDefenders { get; set; }
-
-        public int NumberMidfielders { get; set; }
-
-        public int NumberForwards { get; set; }
-    }
 }
