+++ Audacia.OrToolsPlayground/Examples/PickFplTeam/Models/PickFplTeamOptions.cs
@@ -1,0 +1,16 @@
+namespace Audacia.OrToolsPlayground.Examples.PickFplTeam.Models;
+
+public class PickFplTeamOptions
+{
+    public int MaxPlayersPerTeam { get; set; }
+
+    public int Budget { get; set; }
+
+    public int NumberGoalkeepers { get; set; }
+
+    public int NumberDefenders { get; set; }
+
+    public int NumberMidfielders { get; set; }
+
+    public int NumberForwards { get; set; }
+}
